# bubuka_pro/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микросервиса",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/techs_reqs/{tech_req_id}/": {
            "get": {
                "tags": [
                    "Technicals_Requirements"
                ],
                "summary": "Получение записи",
                "operationId": "38cc04daab69d5ebba75a37ef15baa1e",
                "parameters": [
                    {
                        "name": "tech_req_id",
                        "in": "path",
                        "description": "Id тех. хар-и",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Technical_Requirement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Technicals_Requirements"
                ],
                "summary": "Удаление записи",
                "operationId": "dc4803dab0c5eb09d2f39d40fd27a4c1",
                "parameters": [
                    {
                        "name": "tech_req_id",
                        "in": "path",
                        "description": "Id тех. хар-и",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус ответа",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Technicals_Requirements"
                ],
                "summary": "Изменение записей о технических характеристик",
                "operationId": "4fb2621b7fb8016979b014758c4cd207",
                "parameters": [
                    {
                        "name": "tech_req_id",
                        "in": "path",
                        "description": "Id тех. хар-и",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "os_type": {
                                        "description": "Операционная система",
                                        "type": "string",
                                        "example": "Unix"
                                    },
                                    "specifications": {
                                        "description": "Прочие характеристики",
                                        "type": "string",
                                        "example": "ОЗУ 4 ГБ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Technical_Requirement"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус ответа",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "description": "Ошибки",
                                            "properties": {
                                                "os_type": {
                                                    "type": "string",
                                                    "example": [
                                                        "The os_type field is required when specifications is not present.",
                                                        "The os_type field must be a string.",
                                                        "The os_type field must not be greater than 255 characters."
                                                    ]
                                                },
                                                "specifications": {
                                                    "type": "string",
                                                    "example": [
                                                        "The specifications field is required when os_type is not present.",
                                                        "The specifications field must be a string",
                                                        "The specifications field must not be greater than 255 characters."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "description": "Код ответа",
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/400-Update"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Technical_Requirement": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "body": {
                        "properties": {
                            "technicals_requirements/technical_requirement": {
                                "properties": {
                                    "id": {
                                        "description": "Id записи",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "os_type": {
                                        "description": "Тип операционной системы",
                                        "type": "string",
                                        "example": "Unix"
                                    },
                                    "specifications": {
                                        "description": "Другие характеристики",
                                        "type": "string",
                                        "example": "ОЗУ 16 Гб"
                                    },
                                    "deleted_at": {
                                        "description": "Дата удаления",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "created_at": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    },
                                    "updated_at": {
                                        "description": "Дата изменения",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "400-Update": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Данные идентичны, обновление не требуется"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "405"
                    }
                },
                "type": "object"
            },
            "404": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Запись не найдена"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "404"
                    }
                },
                "type": "object"
            },
...
            "401": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": [
                            "Token not provided",
                            "The token has been blacklisted"
                        ]
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "401"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}